permission

+++ frontend/controllers/StockItemPriceController.php
@@ -8,6 +8,7 @@ use frontend\models\StockItemPriceSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
 
 /**
  * StockItemPriceController implements the CRUD actions for StockItemPrice model.
@@ -17,17 +18,49 @@ class StockItemPriceController extends Controller
     /**
      * @inheritdoc
      */
-    public function behaviors()
-    {
-        return [
-            'verbs' => [
-                'class' => VerbFilter::className(),
-                'actions' => [
-                    'delete' => ['POST'],
-                ],
-            ],
-        ];
-    }
+     public function behaviors()
+     {
+       return [
+       'access' => [
+           'class' => AccessControl::className(),
+           //'only' => ['logout', 'signup'],
+           'rules' => [
+               [
+                   'actions' => ['create', 'index', 'error', 'update', 'view', 'remove','delete','error', 'report', 'long-report'],
+                   'allow'   => true,
+                    'roles'   => ['@'],
+                   'allow'   => true,
+                   'roles'   => ['@'],
+               ],
+           ],
+       ],
+       'verbs'  => [
+           'class'   => VerbFilter::className(),
+           'actions' => [
+               'delete' => ['post'],
+           ],
+       ],
+   ];
+   }
+
+     public function checkPermission()
+     {
+         $userId     = Yii::$app->user->identity->id;
+         $connection = \Yii::$app->db;
+         $checkAdmin = $connection->createCommand("SELECT item_name from auth_assignment WHERE user_id
+             =" . $userId)->queryAll();
+
+             //use this bettween any action you want to give to spasific person like admin
+
+
+         //   if (count($this->checkPermission()) > 0 ) { }
+         return $checkAdmin;
+     }
+
+     public function actionError()
+     {
+         return $this->render('error');
+     }
 
     /**
      * Lists all StockItemPrice models.
@@ -43,6 +76,8 @@ class StockItemPriceController extends Controller
             'dataProvider' => $dataProvider,
         ]);
     }
+
+    
 
     /**
      * Displays a single StockItemPrice model.
@@ -64,6 +99,7 @@ class StockItemPriceController extends Controller
     public function actionCreate()
     {
         $model = new StockItemPrice();
+            if (count($this->checkPermission()) > 0 ) {
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
           Yii::$app->getSession()->setFlash('success',
@@ -75,6 +111,10 @@ class StockItemPriceController extends Controller
                 'model' => $model,
             ]);
         }
+      }else{
+         return $this->redirect(['error']);
+      }
+
     }
 
     /**
@@ -86,6 +126,7 @@ class StockItemPriceController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
+            if (count($this->checkPermission()) > 0 ) {
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
           Yii::$app->getSession()->setFlash('success',
@@ -97,6 +138,10 @@ class StockItemPriceController extends Controller
                 'model' => $model,
             ]);
         }
+      }else{
+         return $this->redirect(['error']);
+      }
+
     }
 
     /**
@@ -107,9 +152,15 @@ class StockItemPriceController extends Controller
      */
     public function actionDelete($id)
     {
+      if (count($this->checkPermission()) > 0 ) {
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+      }else{
+         return $this->redirect(['error']);
+      }
+
     }
 
     /**
